requirements.txt
# --- Core LangChain stack ---
langchain
langchain-core
langchain-community
langchain-experimental
langchain-pinecone
langchain-google-genai

# --- LLM & Embeddings ---
sentence-transformers
google-generativeai
pinecone-client

# --- Utility packages ---
flask
pypdf
python-dotenv

# --- Development ---
-e .

